+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added ESLint linting.
 - Added Prettier formatting.
 - Added Playwright end-to-end testing.
-- Added Testing Library integration testing for React and Vue.
+- Added Testing Library integration testing.
 - Added Vitest unit testing.
 - Added patch-package to remove jsxDev from Angular build.
 - Added Angular integration.